requirements.txt
# MarvelAura2 - Dependencies
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=11.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
setuptools>=70.0.0

# Emotion Detection - Face
# MediaPipe throws errors on Python 3.11/3.12 with strict pins. 
# We use >= and cap dependencies to let pip resolve the best wheel.
mediapipe>=0.10.9
opencv-python-headless>=4.8.0
numpy<2.0.0
protobuf>=3.20.3,<5.0.0

# Emotion Detection - Voice
librosa>=0.10.1
soundfile>=0.12.0

# Speech-to-Text
vosk>=0.3.45
SpeechRecognition>=3.10.0

# Text-to-Speech
edge-tts>=6.1.10

# LLM Integration
openai>=1.14.0
httpx>=0.27.0
google-generativeai>=0.5.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
aiofiles>=23.2.0
